src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gtts==2.4.0
moviepy==1.0.3
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2 